refactor(code): move requests to folders
