+++ app/Http/Controllers/Auth/RedirectController.php
@@ -5,13 +5,13 @@ declare(strict_types=1);
 namespace App\Http\Controllers\Auth;
 
 use App\Http\Controllers\Controller;
-use App\Http\Requests\AuthRequest;
+use App\Http\Requests\Auth\OAuthRequest;
 use Illuminate\Http\RedirectResponse;
 use Laravel\Socialite\Facades\Socialite;
 
 class RedirectController extends Controller
 {
-    public function __invoke(AuthRequest $request): RedirectResponse
+    public function __invoke(OAuthRequest $request): RedirectResponse
     {
         return Socialite::driver($request->provider)->redirect();
     }
