+++ app/Http/Controllers/Settings/Bookmarks/StoreBookmarkController.php
@@ -6,7 +6,7 @@ namespace App\Http\Controllers\Settings\Bookmarks;
 
 use App\Actions\Bookmarks\StoreBookmarkAction;
 use App\Http\Controllers\Controller;
-use App\Http\Requests\StoreBookmarkRequest;
+use App\Http\Requests\Bookmarks\StoreBookmarkRequest;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Session;
