+++ app/Http/Controllers/Settings/Bookmarks/StoreBookmarkGroupController.php
@@ -6,7 +6,7 @@ namespace App\Http\Controllers\Settings\Bookmarks;
 
 use App\Actions\Bookmarks\StoreBookmarkGroupAction;
 use App\Http\Controllers\Controller;
-use App\Http\Requests\StoreBookmarkGroupRequest;
+use App\Http\Requests\Bookmarks\StoreBookmarkGroupRequest;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Session;
