+++ app/Http/Requests/Bookmarks/StoreBookmarkRequest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Bookmarks;
 
-use Auth;
 use Illuminate\Foundation\Http\FormRequest;
+use Illuminate\Support\Facades\Auth;
 
 class StoreBookmarkRequest extends FormRequest
 {
