--- app/Http/Requests/StoreBookmarkGroupRequest.php
@@ -1,24 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Http\Requests;
-
-use Auth;
-use Illuminate\Foundation\Http\FormRequest;
-
-class StoreBookmarkGroupRequest extends FormRequest
-{
-    public function authorize(): bool
-    {
-        return Auth::check();
-    }
-
-    public function rules(): array
-    {
-        return [
-            'name' => 'required',
-            'team_ids' => 'array',
-        ];
-    }
-}
