+++ app/Http/Requests/Teams/AddUserRequest.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace App\Http\Requests\Teams;
 
 use App\Models\User;
-use Auth;
 use Illuminate\Foundation\Http\FormRequest;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Validation\Rule;
 
 class AddUserRequest extends FormRequest
