+++ app/Http/Requests/Teams/StoreTeamRequest.php
@@ -2,23 +2,22 @@
 
 declare(strict_types=1);
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Teams;
 
 use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Support\Facades\Auth;
 
-/* @property-read string $provider */
-class AuthRequest extends FormRequest
+class StoreTeamRequest extends FormRequest
 {
     public function authorize(): bool
     {
-        return Auth::guest();
+        return Auth::check();
     }
 
     public function rules(): array
     {
         return [
-            'provider' => ['in:github'],
+            'name' => 'required|string',
         ];
     }
 }
